Add new fetch bill runs service (#921)
https://eaflood.atlassian.net/browse/WATER-4445

> Part of a series of changes to migrate the legacy view bill runs page into this project

In order to view the bill runs we need to fetch them first! This change adds a new service that generates a summary for each bill run for use in a view bill runs page we'll be adding later.

A key thing to note about this change is it is the first time we've added an [Objection.js paging query](https://vincit.github.io/objection.js/recipes/paging.html#paging). This service is going to support us adding pagination to our pages for the first time via the [GOV.UK pagination component](https://design-system.service.gov.uk/components/pagination/).