Add new check 'busy' bill runs service (#922)
https://eaflood.atlassian.net/browse/WATER-4445

> Part of a series of changes to migrate the legacy view bill runs page into this project

In the legacy service, we made a change to display a notification banner if a bill run is building or cancelling. But it only works if one of the bill runs on the page being viewed is in that state. Select a page of results that doesn't have a building bill run and the banner goes away.

We were limited because we didn't want to break anything with the existing functionality or invest too much in adding new stuff. Now we're migrating the view to this service we have the opportunity to do it right.

This service will be used by another we're adding in a subsequent change to determine which banner to show if any.