Manage 2PT Status on Licence review page (#918)
https://eaflood.atlassian.net/browse/WATER-4371

Due to the complexities of errors being recorded against both returns and charge elements, it has been decided that the user will need to set any licence in review to a ready status to enable the bill run to continue.  This will be achieved through a button on the licence review page.

The button will set the Licence status from review to ready, or vice versa, it will not update or overwrite any of the statuses held against the elements, or the returns.

There is also a highlighted section on the licence review page to show if elements or returns have a review status against them.

This PR will create that button and the associated functionality.