Make path dynamic in paginator
Thanks to a great spot by @Jozzey we realised that the paginator, though intended for reuse, had hard code the links to be `/system/bill-runs`.

If other pages are to be able to reuse it that path must be provided as an arg.

This change updates the presenter. The tests will fail but we wanted to confirm that before fixing them.