Refactor review licence POST pattern
https://eaflood.atlassian.net/browse/WATER-4447

We recently added [Hapi yar session manager to the project](https://github.com/DEFRA/water-abstraction-system/pull/926). This was specifically to support flash notification banners (ones that appear on first GET request but then disappear).

The Review licence screen has a number of these already. Ones for when you toggle the progress and and others for toggling the status. These have been achieved by breaking the [post-redirect-get pattern](https://www.geeksforgeeks.org/post-redirect-get-prg-design-pattern/) and responding directly from the POST request.

As a first step to switching to using **yar** we need to refactor `ReviewLicenceService` into two; `ReviewLicenceService` and `SubmitReviewLicenceService`. This means it will also follow the pattern we've been using in other areas of the project.

This change is that first step. We'll switch to **yar** in the next one.