Returns required journey - Notes page (#906)
https://eaflood.atlassian.net/browse/WATER-4269

This branch adds notes feature to add a note to the return requirement.

The page has a text entry box for notes and a character count below it to display how many characters have been entered. Validation ensures no more than 500 characters are entered.

The content will be saved in the session until the user submits the application.