Arrgh! Support cancelling bill runs
The feedback about checking we were generating the href correctly for bill runs with a 'review' status reminded us we were not properly handling bill runs with a 'cancel' status.

No link should be displayed for these (there is a good chance the bill run will be gone by the time someone clicks the link).

This change covers that omission.