Refactor ReviewLicenceService for GET only
We've split off the functionality that gets used when a POST request is made into `SubmitReviewLicenceService`.

So, now this service only needs to handle when a GET request is made.

On the tests, it was previously understandable that `ReviewLicencePresenter` was stubbed. There is _a lot_ of test data needed to make it work otherwise. But stubbing it is the exception. In all our other tests we've not needed to. We think this is a 'smell' that both the fetch service and the presenter are doing too much. Having to cater for this in the tests is a handy way of highlighting this.

It is beyond the scope of this PR to revisit the service and the presenter. So, for now we limit the change to drop the stubbing and take the data hit. But we'll endeavour to revisit these modules in the future.