Create SubmitReviewLicenceService for POST
The keen eyed will see this is an exact copy of `ReviewLicenceService` before we refactored it!

The scope of this PR is to split the GET and POST request handling to get us in a position we can re-write how the POST is handled to use [yar](https://github.com/DEFRA/water-abstraction-system/pull/926) and fix it to support post-redirect-get again.

So, we not only just copy the code across, we also copy the test as it was.

We promise we'll be updating it in a subsequent change! ;-)