Add charge reference links to licence review page
This PR is an enhancement to the licence review page for two-part tariff bill runs. It is adding a dynamic link to the charge references being displayed. The link will display 1 of two pieces of text depending on if the charge reference has either an aggregate value or a charge adjustment factor. If it does it will display 'Change details' and if it doesn't then 'view details'.