Refactor the submit service to use yar
Now we are setting the banner in the cookie session with the intention of redirecting from the controller we no longer need to fetch the review licence or present it.

We still have to determine which button was the trigger, and from that what needs to be updated and set as a the banner message. But that is all the service has to do now.