Fix validation messages for two-part tariff billable returns page
https://eaflood.atlassian.net/browse/WATER-4370

During testing for the above ticket, it was found that a new validation was needed for the custom quantity that a user can enter. This quantity can not be greater then the authorised quantity on the charge element. This PR is to add this validation in and to amend the original validation messages to update them as per the design teams request.