Add supp. billing notification to view licence (#943)
* Add supp. billing notification to view licence

https://eaflood.atlassian.net/browse/WATER-4323

The legacy view licence page we are replacing displays a notification banner if the licence is flagged for supplementary billing. It could be flagged for PRESROC, SROC or both so the message has to reflect each possible state.

So, this means we need to do the same thing in our version of the view licence page.