Update the two-part tariff review tables to include adjustments (#942)
As part of the work on the two-part tariff review pages we getting ready to add new links to review and change the adjustment factors on a charge reference. In preparation for this change the review tables we currently have need to be updated to accommodate these new values. In this preparation we have decided that the column names for the original source data should stay as they are and any new values that are getting persisted should have the word 'amended' in front of the colum name. For example aggregate and amendedAggregate. This PR is for the table migrations and to clear up some previous code that used the column names the wrong way round.