Add two-part tariff charge reference details review page
As part of the ongoing work for the two-part tariff review screens, this PR introduces the new details page for a charge reference. The charge references have difference charge factors that are displayed on these screens and with future logic can be changed/updated. This initial PR is just the page with its content and no update buttons.