Fix broken PRESROC review bill run links
https://eaflood.atlassian.net/browse/WATER-4445

In [Migrate view bill runs page from legacy UI](https://github.com/DEFRA/water-abstraction-system/pull/925) we were able to implement a new version of the bill runs page!

But our QA team have spotted that the links to PRESROC bill runs that are in `review` are broken. The links we generate are for our review page when they should be for the legacy one.

This change fixes the issue.