Add validation for amend billable returns page (#947)
https://eaflood.atlassian.net/browse/WATER-4370

During the testing of the amend billable returns volume page it was highlighted that there was 1 validation test missing. When a user enters a new volume for the billable returns, the volume currently cannot be greater then the authorised volume on the charge reference. This is correct however it should also not go above the charge reference authorised volume as well. So whichever volume is LOWER the authorised charge element or the authorised charge reference. This PR is to add in that additional validation.