Saving input from return requirements in session (#939)
* Saving input from return requirements in session

During the initial development of the return requirements pages, user input
was not being saved in session. Later tickets and pages have this functionality in place.
The session data is needed for the Check Your Answers page at the end of the journey.

This PR is focused on saving user input into the session where it is missing.