Replay submitted data with a validation error
When a user inputs data that returns a validation error, the page will clear the input.
This needs to be fixed so that data with errors is replayed to the user so they can make
changes to what they have submitted.

This PR is focused on replaying submitted data with a validation error.