Refactor to load tabs as pages in view licence (#953)
* View License page refactor to load pages mapped to controllers when navigating by tabs.

This PR maps the generic controller view license controller to the summary tabs and associated data.

To switch using hrefs instead of anchor tabs the default GDS javascript is turned off and the html used instead of the component (this only supports anchor tags)

https://eaflood.atlassian.net/browse/WATER-4457

Context:

The current view licence page is built using tabs. The design system makes it very clear that tabs should only be used to allow users to switch between related pieces of information. The example it gives is cases (from a case management system) split over Past day, Past week, Past month and Past year.

They very clearly state

Do not use the tabs component if the total amount of content the tabs contain will make the page slow to load. For this reason, do not use the tabs component as a form of page navigation.

We may not have meant to start out using them in this way. But we’ve ended up there. And because of the issue of performance, we’ve made design decisions based on negating the issue rather than what users need.

For example, we show only the first 10 returns on the Returns tab and then a link to view the full list. This is because the tab component works on the basis that the data for all the tabs is fetched in one go, and the tabs are all generated on the first request. Then HTML magic switches what HTML panel the user sees when they click a tab.

Redesigning the view licence page is beyond the scope of the refactoring. But we can take steps to break away from the need to fetch and generate the tabs all in one go.