Fix amended billable returns banner two-part tariff review
During testing it was noted that the amended billable returns banner broke during a previous PR (https://github.com/DEFRA/water-abstraction-system/pull/942). It is currently always being displayed. This is because it displays the banner if the amended volume doesn't equal the source volume and since the 'calculated' column no longer exists in the db it therefore never equals each other. This change is to fix the banner to display it only when there is a change to the amended amount.