Add .editorconfig to keep project files consistent (#960)
https://github.com/DEFRA/water-abstraction-team/issues/115

We've been fortunate enough to grow as a team this year which means we've had several new developers join us! 🥳

Unfortunately, it has exposed our onboarding still leaves a lot to be desired. We find we keep having to highlight the same coding convention or stylistic issues in new members' initial PRs. This is irrespective of whether we have written up the convention or not.

Wouldn't it be great if we used all this computing power we interact with daily to help them get it right the first time? That we help them fall into the [pit of success](https://www.hackterms.com/pit%20of%20success)?

One idea is to add an [.editorconfig](https://editorconfig.org/) to the repo. Depending on folks IDE choice it will either be recognised out-of-the-box or they'll need an extension. 

Unlike other tools having the file doesn't do anything _directly_ to your project. Instead, it tells your chosen editor how to behave when working with the files. For example, when you indent something whether to use tabs or spaces and how many. If your editor can remove whitespace on save, the `.editorconfig` will tell it whether to do that or not.

Even for folks using an incompatible editor or who choose not to install the extension, we at least have something that defines what we expect and can be used in our CI pipeline.

This change adds an initial version generated using the [VSCode extension](https://marketplace.visualstudio.com/items?itemName=EditorConfig.EditorConfig) based on current editor settings.

It also updates our GitHub CI workflow to check that pushed changes match our conventions.

---

This change highlighted that the line endings in our migrations were inconsistent. So, the fix for those is included.