Persist charge reference authorised volume for two-part tariff review (#966)
https://eaflood.atlassian.net/browse/WATER-4435
https://eaflood.atlassian.net/browse/WATER-4440

As part of the work on our two-part tariff review screens we are now building new pages to edit the authorised volume for the charge reference. This PR is the prep work for those pages, adding the charge reference authorised volume into our persisted data and an amended column for it as well. This PR also covers a content change on the review licence page, where authorised volume is shown, but is currently worked out by adding up the totals of the authorised volumes from the charge elements. This is being changed to the authorised volume on the charge reference instead.