Check your answers page with notes for additions, changes and deletions (#923)
https://eaflood.atlassian.net/browse/WATER-4291

This feature allows the user to see notifications on the page when a note is added, changed or deleted.

A delete-note service has been added which is accessed via a simple GET request. This will delete any note in the session attached to the requirement.

Notifications are shown using hapi/yar and are immediately removed upon first view on the page.