Remove Licence from bill 2PT bill run during review (#927)
https://eaflood.atlassian.net/browse/WATER-4216

On the licence review page add a button for “remove from bill run”. When clicked you should be presented with a confirmation page with a confirm button, when this is clicked the licence should be removed from the current bill run.

Once the licence is removed the user should be returned to the bill run review page with a banner confirming the licence has been removed. When removed the licence will be flagged for two-part tariff supplementary billing.