Alter and add migration for amendAllocated column two-part tariff review
Oppsie, I messed up! During testing for the two-part tariff review screens it was noted that some of the data for a charge elements allocated volume wasn't persisting correctly. This is due to the table's column not being the correct data type. The initial migration for the creation of the column calculated (now called amended_allocated) was on April 17th. The column wasn't created properly and needed to be amended so that the decimal points weren't getting cut off. Having checked the commit history in the repo there was no tags at the time, so rather than create a new migration, I went and edited the original migration instead. This is where my oppsie happened. The edited migration commit happened on the 17th (before the repo got taggged for release) however due to the length of PR I was working on the actual PR itself didn't get merged until the 24th. The repo got tagged for a release between this on the 19th. This means the initial migration has been run on production and the edited version of the migration will not work for updating the column data type. This PR is reverting the migration file back to its original state and add a new migration that can be run to update the column data type. *facePalm*