Add support for multiple return reqs. during setup
https://eaflood.atlassian.net/browse/WATER-4463

The first problem we had to tackle was how we hold the selections a user is making on _any_ journey in the app till they are ready to commit them.

With that resolved (welcome the `SessionModel`), we got on with building the manual 'returns required' and 'no returns required' journeys for the return requirements feature. That meant saving and then replaying what a user had selected so we could get to the 'Check your answers' page with something for the QA team to look at.

Now we are on the cusp of building the return requirement summary card for 'Check your answers'. But hot on the heels of that will be the need to allow users to add additional return requirements! We ..cough.. haven't catered for that in our session handling and the journey 😰.

The key things we need to be able to do on any page in the journey is

- know which requirement we are extracting a value from for presenting in the page
- know which requirement we are updating when a user submits a new value

This change implements that support.