Rename add-note route in ret. requirements to note
https://eaflood.atlassian.net/browse/WATER-4261

The `/add-note` route on the return requirements setup journey is inconsistent with the rest of our routes because it is the only one that includes a verb. This really becomes a problem when we have to go to the same endpoint to change an existing note; it no longer makes sense.

So, this change is some housekeeping to get the route name inline with the others and use simply `/note`.