Correct validators test folder name
https://eaflood.atlassian.net/browse/WATER-4261

When we first start creating validators for the returns requirement setup we focused on the 'no-returns-required' journey. It is the shortest and easiest!

But we clearly got it lodge in our minds a little too much as we named the folder after it rather than the overall feature. Doh!

This change corrects that.