Correct validators test folder name (#979)
https://eaflood.atlassian.net/browse/WATER-4261

When we first started creating validators for the returns requirement setup we focused on the 'no-returns-required' journey. It is the shortest and easiest!

But we must have got it lodged in our minds a little _too_ much as we named the folder after it rather than the overall feature. Doh!

This change corrects that.