Fix broken delete & check answers service tests
The tests were setting an ID property in `data` which was breaking the changes we had made. Obviously, when fetched this will be elevated and overwrite the existing ID property causing problems.

Fortunately, the property should never have been set in these tests. We never use it in the tests and we never set a `session.data.id` property anywhere in the journey anyway!