Enhance the SessionModel to expose and update data (#981)
https://eaflood.atlassian.net/browse/WATER-4463

We created our `SessionModel` and its backing table out of the need to be able to store data temporarily, for example, when a user is following a journey to create a new return requirement.

We don't want to use cookies, because their size is limited, and they have a habit of sticking around and polluting new sessions. The 'proper' way is with a server-side session cookie backed by Redis. But we're trying to simplify our service not embed the existing complexity.

We have PostgreSQL, reading and writing to it is fast enough and scalable enough for our use case so why not use it!?

So, that is what we did. We made it simple and flexible by having just one property, `data`, which is a JSONB field so can hold whatever data we need to support whatever journey.

Currently, it's used in the bill run and return requirements set-up journeys and will no doubt be used for more in the future.

What we have seen though are some common patterns and duplications emerging. The services we've created to support these journeys are

- fetching the session
- extracting the properties they need via the `session.data` instance
- saving changes by first extracting a copy of the existing `data` property, updating it, then calling `patch({ data: currentData })` on the instance

With these changes, we intend to make things a little slicker!

Using the [Objection.js $afterFind() hook](https://vincit.github.io/objection.js/api/model/instance-methods.html#afterfind) we can automatically elevate the properties in `data` to the top-level. This means we can use, for example, `session.siteDescription` instead of `session.data.siteDescription`. It also means new properties can be added at the top level. No more forgetting to add `data` and wondering why it isn't working!

When it comes to updating, we can simplify things for callers. We can move the work of updating the value of `data` into the model itself. If we have a session instance, which we do in all our current `Submit` services. Having set your new property, for example, `session.reason = 'major-change'`, you can then just call `session.update()`. The model will handle ensuring the new property you set is applied to `data` along with all its existing values. It works the same for properties that already existed, which were elevated to the top level during the find and then updated. `update()` will ensure the new value is saved.