Update rtn. req. setup journey to use new session (#984)
https://eaflood.atlassian.net/browse/WATER-4463

We created our `SessionModel` and its backing table out of the need to be able to store data temporarily, for example, when a user is following a journey to create a new return requirement.

After using it in our bill run setup and return requirement journeys we spotted there were some improvements we could make. So, we [Enhanced the SessionModel to expose and update data](https://github.com/DEFRA/water-abstraction-system/pull/981).

The next step is to update the existing journeys to take advantage of these enhancements.

In this change, we update the return requirement setup journey.