Implement initial multi-requirements support
This is a first pass at the solution. We initialise the session with a new array of requirements containing a blank object.

When we get into the journey we add an index to the path which tells the steps in the journey which requirement we are currently setting up.

Unfortunately this change is messier than intended. The boy scout rule kicked in and we couldn't help ourselves! So, along with these changes we have also done the following.

- Fix back link functionality and move the logic into the presenters. Now the pages not only know to continue to the check your answers page, but go back there as well if the back link is clicked
- remove references to payload in a number of the presenter comments
- make the JSDoc comments consistent
- remove adding `checkYourAnswers` as a property when pages error. It is not needed
- alignment issues in the views
- add double-click prevention to all buttons

We haven't gone through everything in the journey in this fashion but we'll endeavour to before the change is complete. This is also a WIP commit. We are aware that this will break all the related tests! But it does work!