Exclude long strings from ESLint max-len rule
https://github.com/DEFRA/water-abstraction-team/issues/115

Whilst we should be making every effort to keep lines shorter than 120 characters the intent for the rule is to stop folks from stringing together complex one-liners.

There are times we need to include long pieces of text, build strings from verbose variable names or reference URLS that will cause us to go over the limit.

We don't want to make them more complex just for the sake of a linter. So, we add this config to exclude these types from the rule.