Add View Licence contact details page
https://eaflood.atlassian.net/browse/WATER-4433
https://eaflood.atlassian.net/browse/WATER-4434

The existing service handling view licence is slow because it loads all the data for the tabs in one render. Work has been done previously to refactor the summary page to load only the summary information.

This change will introduce a contact details controller, service and presenter to handle the view licence contact details page.

This will share the same view as the summary page and load the same 'common data' established in [previous work](#957).